Properly format as extension
