+++ Project.toml
@@ -13,10 +13,12 @@ Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
 
 [weakdeps]
 MCMCDiagnosticTools = "be115224-59cd-429b-ad48-344e309966f0"
+NCDatasets = "85f8d34a-cbdd-5861-8df4-14fed0d494ab"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 
 [extensions]
 InferenceObjectsMCMCDiagnosticToolsExt = ["MCMCDiagnosticTools", "Random"]
+InferenceObjectsNCDatasetsExt = ["DimensionalData", "NCDatasets"]
 
 [compat]
 Compat = "3.46.0, 4.2.0"
@@ -24,6 +26,7 @@ DimensionalData = "0.24"
 EvoTrees = "0.14"
 MCMCDiagnosticTools = "0.3.4"
 MLJBase = "0.21"
+NCDatasets = "0.12"
 OffsetArrays = "1"
 OrderedCollections = "1"
 Requires = "0.5, 1"
@@ -34,6 +37,7 @@ julia = "1.6"
 EvoTrees = "f6006082-12f8-11e9-0c9c-0d5d367ab1e5"
 MCMCDiagnosticTools = "be115224-59cd-429b-ad48-344e309966f0"
 MLJBase = "a7f614a8-145f-11e9-1d2a-a57a1082229d"
+NCDatasets = "85f8d34a-cbdd-5861-8df4-14fed0d494ab"
 OffsetArrays = "6fe1bfb0-de20-5000-8ca7-80f57d26f881"
 OrderedCollections = "bac558e1-5e72-5ebc-8fee-abe8a469f55d"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
@@ -41,4 +45,4 @@ Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["EvoTrees", "MCMCDiagnosticTools", "MLJBase", "OffsetArrays", "OrderedCollections", "Random", "Statistics", "Test"]
+test = ["EvoTrees", "MCMCDiagnosticTools", "MLJBase", "NCDatasets", "OffsetArrays", "OrderedCollections", "Random", "Statistics", "Test"]
