+++ src/InferenceObjects.jl
@@ -43,6 +43,7 @@ include("convert_dataset.jl")
 include("convert_inference_data.jl")
 include("from_namedtuple.jl")
 include("from_dict.jl")
+include("io.jl")
 
 if !EXTENSIONS_SUPPORTED
     using Requires: @require
